Extract getActionItems to utils
